+++ app/app.routes.ts
@@ -2,8 +2,8 @@
 import { Routes } from "@angular/router";
 
 import { NavigationGuard } from "app/components/base/guards";
+import { JobScheduleHomeComponent } from "app/components/job-schedule/home/job-schedule-home.component";
 import { JobGraphsComponent } from "app/components/job/graphs/job-graphs-home";
-import { JobScheduleHomeComponent } from "app/components/jobSchedule/home/job-schedule-home.component";
 import { ChooseActionComponent } from "app/components/market/application-action";
 import { LocalTemplateBrowserComponent } from "app/components/market/local-template-browser";
 import { SubmitLocalTemplateComponent } from "app/components/market/submit-local-template";
