+++ app/components/job-schedule/action/delete/index.ts
