+++ app/components/job-schedule/browse/job-schedule-list.component.ts
